Los service para sale estan listos y el metodo de listar ya esta listo

+++ src/main/java/bunker/dao/SaleDao.java
@@ -3,6 +3,6 @@ package bunker.dao;
 import bunker.entidad.Sale;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SaleDao extends JpaRepository<Sale, Long>
+public interface SaleDao extends JpaRepository<Sale, Integer>
 {
 }
